Removed interface InTrayManager
